+++ CHANGELOG.md
@@ -1,10 +1,11 @@
 # CHANGELOG
 
 ## 0.9.11 - 07.04.2023
-- fix dependencies inclusion
+- fix dependencies inclusion - see [#20](https://github.com/openTdataCH/ojp-js/pull/20)
   - avoid usage of `allowSyntheticDefaultImports` by using `import * as`
   - declare `GeoJSON` types
   - no need to import the heavy mapbox-gl lib, just need the TS `@types`
+- adds default stage to be used without registering a new key - see [#21](https://github.com/openTdataCH/ojp-js/pull/21)
 
 ## 0.9.10 - 02.04.2023
 - fix `xmldom` inclusion - see [#19](https://github.com/openTdataCH/ojp-js/pull/19)